+++ docs/changelog.md
@@ -21,6 +21,7 @@ Release is mostly focused on tidying up the docs and code base and on resource c
 * Added `--memberlist.address`, `--memberlist.advertise-address` and  `--memberlist.members` flags to configure cluster registry.
 * Added a raft cluster snapshot mode option `--raft.snapshot-recovery-type` which defaults to new `checkpoint` mode.
 * Added a flag for configuring shared table cache size `--storage.table-cache-size` which defaults to `1024`.
+* Added support for `gzip`, and `snappy` compression for API calls.
 
 ### Improvements
 * Compressed Log replication messages to lower API bandwidth usage.